% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_distribution.R
\name{create_PSD}
\alias{create_PSD}
\title{Create particle size distribution}
\usage{
create_PSD(db, vct.dir, quantile = 50, param = "Qc_mid", breaks)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct.dir}{VCT file directory..}

\item{quantile}{Filtering function.}

\item{param}{Parameter from VCT to create the size distribution.
Can be either diameter (diam_lwr, diam_mid or diam_upr)
or carbon quotas (Qc_lwr, Qc_mid or Qc_upr)}

\item{breaks}{Breaks must be a vector of values defining the breaks for the size distribution.}
}
\value{
Size distribution
}
\description{
Create particle size distribution
}
\examples{
\dontrun{
 
breaks <- 'something here'
distribution <- create_PSD(db, vct.dir, quantile=50, channel="Qc_mid", breaks)
}
}
