% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.functions_model.R
\name{delta}
\alias{delta}
\title{Delta function. Fraction of cells that divide between t and t + dt}
\usage{
delta(volbins, dmax, b)
}
\arguments{
\item{volbins}{volbins}

\item{dmax}{dmax}

\item{b}{b}
}
\value{
delta
}
\description{
Delta function. Fraction of cells that divide between t and t + dt
}
