% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.functions_model.R
\name{determine_opt_para}
\alias{determine_opt_para}
\title{model optimization usin Diff Evolution (require DEoptim package)}
\usage{
determine_opt_para(distribution = distribution, Edata = Edata,
  resol = resol)
}
\arguments{
\item{distribution}{blabla}

\item{Edata}{blabla}

\item{resol}{blabla}
}
\value{
optimization
}
\description{
model optimization usin Diff Evolution (require DEoptim package)
}
\examples{
\dontrun{
}
}
