% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.functions_model.R
\name{gamma_t}
\alias{gamma_t}
\title{Gamma function. Fraction of cells that grow into next size class between t and t + dt}
\usage{
gamma_t(Edata, gmax, E_star)
}
\arguments{
\item{Edata}{Edata}

\item{gmax}{gmax}

\item{E_star}{E_star}
}
\value{
gamma
}
\description{
Gamma function. Fraction of cells that grow into next size class between t and t + dt
}
