% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{matrix_conct_fast}
\alias{matrix_conct_fast}
\title{Construct matrix A(t) for each time step within an hour based on delta and gamma at each 10 minute time #' intervals then construct B(t) which is A(t)'s multiplied for the given hour
multiply B(t)*w(t) to get the projection to next hour dist if desired}
\usage{
matrix_conct_fast(hr, Edata, volbins, gmax, dmax, b, E_star, resol)
}
\arguments{
\item{hr}{blabla}

\item{Edata}{blabla}

\item{volbins}{blabla}

\item{gmax}{blabla}

\item{dmax}{blabla}

\item{b}{blabla}

\item{E_star}{blabla}

\item{resol}{blabla}

\item{breaks}{blabla}
}
\value{
Matrix B(t)
}
\description{
Construct matrix A(t) for each time step within an hour based on delta and gamma at each 10 minute time #' intervals then construct B(t) which is A(t)'s multiplied for the given hour
multiply B(t)*w(t) to get the projection to next hour dist if desired
}
\examples{
\dontrun{
}
}
