% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_distribution.R
\name{plot_PSD}
\alias{plot_PSD}
\title{Plot size distribution}
\arguments{
\item{distribution}{Data frame of size distribution over time, (x time; y size classes). 
First column must be time (POSIXt class object), second column must name of the population;
other columns represent the different size classes.
Size classes can represent either diameter or carbon quota (assuming spherical particles).}

\item{lwd}{Line width for the lines}

\item{z.type}{"lin" for linear scaling of z values, "log" for logarithmic scaling}
}
\value{
Plot carbon biomass in each size class
}
\description{
Plot size distribution
}
\examples{
\dontrun{
plot_PSD(distribution)
}
}
