% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{rho_t}
\alias{rho_t}
\title{Rho function. Fraction of cells that shrink between t and t + dt
Assumptions:
1) rate of respiration is a function of photosynthetic rate (gamma function)
2) rate of respiration constant (h-1) constant over the nighttime period 
(polysaccharide is drawn down linearly over the nighttime period)}
\usage{
rho_t(y)
}
\arguments{
\item{y}{Gamma values}
}
\value{
rho
}
\description{
Rho function. Fraction of cells that shrink between t and t + dt
Assumptions:
1) rate of respiration is a function of photosynthetic rate (gamma function)
2) rate of respiration constant (h-1) constant over the nighttime period 
(polysaccharide is drawn down linearly over the nighttime period)
}
