% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{sigma_hl}
\alias{sigma_hl}
\title{This function calculates the sum of squares of the of the differences between the hourly observations 
and the model given the specified parameters using Hubert Loss (1954) approach.}
\usage{
sigma_hl(
  params = params,
  Edata = Edata,
  distribution = distribution,
  resol = resol
)
}
\arguments{
\item{params}{blabla}

\item{Edata}{blabla}

\item{distribution}{blabla}

\item{resol}{blabla}
}
\value{
goodness of fit
}
\description{
This function calculates the sum of squares of the of the differences between the hourly observations 
and the model given the specified parameters using Hubert Loss (1954) approach.
}
\examples{
\dontrun{
}
}
