% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{sigma_lsq}
\alias{sigma_lsq}
\title{This function calculates the sum of squares of the of the differences between the hourly observations and 
the model given the specified parameters
This function returns a column vector - called by "determine.opt.para" for the optimization.}
\usage{
sigma_lsq(
  params = params,
  Edata = Edata,
  distribution = distribution,
  resol = resol
)
}
\arguments{
\item{params}{blabla}

\item{Edata}{blabla}

\item{distribution}{blabla}

\item{resol}{blabla}
}
\value{
goodness of fit
}
\description{
This function calculates the sum of squares of the of the differences between the hourly observations and 
the model given the specified parameters
This function returns a column vector - called by "determine.opt.para" for the optimization.
}
\examples{
\dontrun{
}
}
