% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.size_distribution.R
\name{size.distribution}
\alias{size.distribution}
\title{Create particle size distribution}
\usage{
size.distribution(db, vct.dir, quantile = 50, param = "Qc_mid", breaks)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct.dir}{VCT file directory..}

\item{quantile}{Filtering function.}

\item{param}{Parameter from VCT to create the size distribution.
Can be either diameter (diam_lwr, diam_mid or diam_upr)
or carbon quotas (Qc_lwr, Qc_mid or Qc_upr)}

\item{breaks}{Breaks must be a sequence defining the breaks for the size distribution.}
}
\value{
Size distribution
}
\description{
Create particle size distribution
}
\examples{
\dontrun{
distribution <- size.distribution(db, vct.dir, quantile=50, popname="synecho", channel="Qc_mid", delta=0.125, m=60)
}
}
