% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.size_distribution.R
\name{transform_PSD}
\alias{transform_PSD}
\title{Manipulate size distribution}
\arguments{
\item{distribution}{Data frame of size distribution over time, (x time; y size classes). 
First column must be time (POSIXt class object), second column must name of the population; other columns represent the different size classes.
Size classes can represent either diameter or carbon quota (assuming spherical particles).}

\item{time.step}{Time resolution (must be higher than 3 minutes). Default is 1 hour}

\item{diam.to.Qc}{Convert diameters into carbon quotas as described in
Menden-Deuer, S. and Lessard, E. J. Carbon to volume relationships for dinoflagellates, diatoms, and other protist plankton.
Limnol. Oceanogr. 45, 569–579 (2000).}

\item{Qc.to.diam}{Convert carbon quotas into diameters (reciprocal of diam.to.Qc)}

\item{abundance.to.biomass}{Calcualte total carbon biomass in each size class (abundance x Qc). 
Warning: If size class values represent diameters, make sure to set diam.to.Qc = TRUE.}

\item{size.interval.to.mean}{Transform size class intervals to mean values (i.e. convert breaks (min, max] to mean).}
}
\value{
Size distribution
}
\description{
Manipulate size distribution
}
\examples{
\dontrun{
distribution <- transform_PSD(distribution, time.step="1 hour")
}
}
